JaeunLee 1시간 전 편집됨
KY 에서는 Set 함수로 tm 을 바로 초기화하는데,

Eigen 에서 transform matrix(4,4)에 바로 초기화 시켜도 되는것인지,
affline rx,ry,rz 만들고 각각의 곱을 matrix 화 시키고 translation 까지 해줘야 하는것인지 ?

KyGeoICP .cpp - 280 수정중

+++ Kernel/Base/KyTrans.h
@@ -111,6 +111,6 @@ public:
 
 public:
 	static void MovePts(Eigen::Matrix4d tMat, std::vector<Eigen::Vector3d>& arrPts);
-	static void MovePts(Eigen::Matrix4d tMat, std::vector<Eigen::Vector3f>& arrPts);
+	static void MovePts(Eigen::Matrix4f tMat, std::vector<Eigen::Vector3f>& arrPts);
 	static void MovePts(Eigen::Matrix4d tMat, std::vector<std::vector<Eigen::Vector3d>>& arrPtss);
 };
